Move Old CFG to .Sonar directory and namespace (#4820)
