+++ analyzers/src/SonarAnalyzer.CFG/Sonar/AbstractControlFlowGraphBuilder.cs
@@ -24,7 +24,7 @@ using System.Collections.Immutable;
 using System.Linq;
 using Microsoft.CodeAnalysis;
 
-namespace SonarAnalyzer.ControlFlowGraph
+namespace SonarAnalyzer.CFG.Sonar
 {
     public abstract class AbstractControlFlowGraphBuilder
     {
