+++ analyzers/src/SonarAnalyzer.CFG/Sonar/Blocks/JumpBlock.cs
@@ -20,33 +20,40 @@
 
 using System;
 using System.Collections.Generic;
-using System.Collections.Immutable;
 using Microsoft.CodeAnalysis;
 
-namespace SonarAnalyzer.ControlFlowGraph
+namespace SonarAnalyzer.CFG.Sonar
 {
-    public class BranchBlock : Block
+    public sealed class JumpBlock : SimpleBlock
     {
-        internal BranchBlock(SyntaxNode branchingNode, params Block[] successors)
+        internal JumpBlock(SyntaxNode jumpNode, Block successor, Block wouldBeSuccessor)
+            : base(successor)
         {
-            this.successors = successors ?? throw new ArgumentNullException(nameof(successors));
-            BranchingNode = branchingNode ?? throw new ArgumentNullException(nameof(branchingNode));
+            JumpNode = jumpNode ?? throw new ArgumentNullException(nameof(jumpNode));
+            WouldBeSuccessor = wouldBeSuccessor;
         }
 
-        public SyntaxNode BranchingNode { get; }
+        public SyntaxNode JumpNode { get; }
 
-        protected readonly Block[] successors;
+        /// <summary>
+        /// If there was no jump, this block would be the successor.
+        /// It can be null, when it doesn't make sense. For example in case of lock statements.
+        /// </summary>
+        public Block WouldBeSuccessor { get; private set; }
 
-        public override IReadOnlyList<Block> SuccessorBlocks => ImmutableArray.Create(this.successors);
+        internal override Block GetPossibleNonEmptySuccessorBlock()
+        {
+            // JumpBlock can't be removed by the CFG simplification, unlike the base class SimpleBlock
+            return this;
+        }
 
         internal override void ReplaceSuccessors(Dictionary<Block, Block> replacementMapping)
         {
-            for (var i = 0; i < this.successors.Length; i++)
+            base.ReplaceSuccessors(replacementMapping);
+
+            if (WouldBeSuccessor != null && replacementMapping.ContainsKey(WouldBeSuccessor))
             {
-                if (replacementMapping.ContainsKey(this.successors[i]))
-                {
-                    this.successors[i] = replacementMapping[this.successors[i]];
-                }
+                WouldBeSuccessor = replacementMapping[WouldBeSuccessor];
             }
         }
     }
