+++ analyzers/src/SonarAnalyzer.CFG/Sonar/Blocks/UsingEndBlock.cs
@@ -25,7 +25,7 @@ using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
 using SonarAnalyzer.CFG.Helpers;
 
-namespace SonarAnalyzer.ControlFlowGraph.CSharp
+namespace SonarAnalyzer.CFG.Sonar
 {
     public class UsingEndBlock : SimpleBlock
     {
