+++ analyzers/src/SonarAnalyzer.CFG/Sonar/CSharpControlFlowGraph.cs
@@ -23,7 +23,7 @@ using System.Diagnostics;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp;
 
-namespace SonarAnalyzer.ControlFlowGraph.CSharp
+namespace SonarAnalyzer.CFG.Sonar
 {
     public static class CSharpControlFlowGraph
     {
