+++ analyzers/tests/SonarAnalyzer.UnitTest/CBDE/MlirTestUtilities.cs
@@ -25,8 +25,7 @@ using System.Linq;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using SonarAnalyzer.CBDE;
-using SonarAnalyzer.ControlFlowGraph;
-using SonarAnalyzer.ControlFlowGraph.CSharp;
+using SonarAnalyzer.CFG.Sonar;
 
 namespace SonarAnalyzer.UnitTest.CBDE
 {
